Message TBD

+++ lib/fsm.c
@@ -78,7 +78,7 @@ static int fsmClose(int *wfdp);
 static char * fsmFsPath(rpmfi fi, const char * suffix)
 {
     const char *bn = rpmfiBN(fi);
-    return rstrscat(NULL, *bn ? bn : "/", suffix ? suffix : "", NULL);
+    return rstrscat(NULL, *bn ? bn : ".", suffix ? suffix : "", NULL);
 }
 
 static int fsmLink(int odirfd, const char *opath, int dirfd, const char *path)
@@ -273,9 +273,22 @@ static int fsmStat(int dirfd, const char *path, int dolstat, struct stat *sb)
     return rc;
 }
 
-static int fsmRmdir(int dirfd, const char *path)
+static int fsmRmdir(int dirfd, rpmfi fi, const char *path)
 {
-    int rc = unlinkat(dirfd, path, AT_REMOVEDIR);
+    int rc = 0;
+
+    if (rstreq(path, ".")) {
+	/* Path-based operation needed to avoid EINVAL */
+	path = rpmfiDN(fi);
+	if (rstreq(path, "/"))
+	    /* Don't even try */
+	    return rc;
+	rc = rmdir(path);
+    } else {
+	/* Normal fd-based operation */
+	rc = unlinkat(dirfd, path, AT_REMOVEDIR);
+    }
+
     if (_fsm_debug)
 	rpmlog(RPMLOG_DEBUG, " %8s (%d %s) %s\n", __func__,
 	       dirfd, path, (rc < 0 ? strerror(errno) : ""));
@@ -542,9 +555,9 @@ static int fsmRename(int odirfd, const char *opath, int dirfd, const char *path)
     return rc;
 }
 
-static int fsmRemove(int dirfd, const char *path, mode_t mode)
+static int fsmRemove(int dirfd, rpmfi fi, const char *path, mode_t mode)
 {
-    return S_ISDIR(mode) ? fsmRmdir(dirfd, path) : fsmUnlink(dirfd, path);
+    return S_ISDIR(mode) ? fsmRmdir(dirfd, fi, path) : fsmUnlink(dirfd, path);
 }
 
 static int fsmChown(int fd, int dirfd, const char *path, mode_t mode, uid_t uid, gid_t gid)
@@ -1093,7 +1106,7 @@ setmeta:
 		continue;
 
 	    if (fp->stage > FILE_NONE && !fp->skip) {
-		(void) fsmRemove(di.dirfd, fp->fpath, fp->sb.st_mode);
+		(void) fsmRemove(di.dirfd, fi, fp->fpath, fp->sb.st_mode);
 	    }
 	}
     }
@@ -1151,7 +1164,8 @@ int rpmPackageFilesRemove(rpmts ts, rpmte te, rpmfiles files,
         if (fp->action == FA_ERASE) {
 	    int missingok = (rpmfiFFlags(fi) & (RPMFILE_MISSINGOK | RPMFILE_GHOST));
 
-	    rc = fsmRemove(di.dirfd, fp->fpath, fp->sb.st_mode);
+	    /* printf(">>> %s%s\n", rpmfiDN(fi), fp->fpath); */
+	    rc = fsmRemove(di.dirfd, fi, fp->fpath, fp->sb.st_mode);
 
 	    /*
 	     * Missing %ghost or %missingok entries are not errors.
